parser/expression: update tests
- removed parser_help_test.go: obsoleted by the testing/utils package
- removed parser_test.go: obsoleted by generated tests
- made template up to date with changes of parsers and symbol tables
  - use testing/utils package
  - TODO: obsolete isbuiltin..
- regenerated tests
